Detect unsupported template and display warning
If the jQuery selector does not match any DOM elements, then most likely
the DokuWiki template does not use the standard markup for Page Tools.

Instead of silently failing (resulting in page deletion without
requesting user confirmation first), we now display a warning in the
browser's console with a link to create a new issue in the tracker.

Fixes #6